% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bhai_pop_est}
\alias{bhai_pop_est}
\title{BHAI Population Estimates}
\format{
A tibble with population estimates by infection and country:
\describe{
\item{infection}{Type of healthcare-associated infection}
\item{cases_point_estimate}{Point estimate for number of cases}
\item{cases_lower_ci}{Lower 95\% confidence interval for cases}
\item{cases_upper_ci}{Upper 95\% confidence interval for cases}
\item{deaths_point_estimate}{Point estimate for number of deaths}
\item{deaths_lower_ci}{Lower 95\% confidence interval for deaths}
\item{deaths_upper_ci}{Upper 95\% confidence interval for deaths}
\item{daly_point_estimate}{Point estimate for Disability-Adjusted Life Years}
\item{daly_lower_ci}{Lower 95\% confidence interval for DALYs}
\item{daly_upper_ci}{Upper 95\% confidence interval for DALYs}
\item{yll_point_estimate}{Point estimate for Years of Life Lost}
\item{yll_lower_ci}{Lower 95\% confidence interval for YLL}
\item{yll_upper_ci}{Upper 95\% confidence interval for YLL}
\item{yld_point_estimate}{Point estimate for Years Lived with Disability}
\item{yld_lower_ci}{Lower 95\% confidence interval for YLD}
\item{yld_upper_ci}{Upper 95\% confidence interval for YLD}
\item{country}{Country or region (Germany or European Union)}
\item{population}{Total Population of the country or region}
}
}
\source{
Derived from BHAI simulations on German PPS 2011 and EU PPS data
}
\usage{
bhai_pop_est
}
\description{
Total population-level estimates of healthcare-associated infection burden
from BHAI analysis. This dataset contains parsed point estimates and 95\%
confidence intervals for cases, deaths, DALYs, YLL, and YLD.
}
\details{
This dataset is created by:
\enumerate{
\item Running BHAI simulations (nsim = 500) on German and EU PPS data
\item Formatting results with \code{bhai.prettyTable()}
\item Parsing formatted strings to extract numeric estimates and confidence intervals
}

All estimates are population-level annual burden estimates.
}
\examples{
data(bhai_pop_est)

# View structure
str(bhai_pop_est)

# Infection with highest mortality
bhai_pop_est |>
  filter(country == "Germany") |>
  arrange(desc(deaths_point_estimate)) |>
  select(infection, deaths_point_estimate, deaths_lower_ci, deaths_upper_ci)
}
\seealso{
\link{bhai_pps_sample_distribution}, \link{bhai_strata_summary}
}
\keyword{datasets}
