% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_pps.R
\name{flatten_pps}
\alias{flatten_pps}
\title{Flatten PPS Data Structure}
\usage{
flatten_pps(pps_obj)
}
\arguments{
\item{pps_obj}{An S4 object of class PPS from the BHAI package}
}
\value{
A tibble with one row per infection type, containing:
\describe{
\item{infection}{Type of healthcare-associated infection}
\item{num_hai_patients}{Number of patients with HAI}
\item{num_survey_patients}{Total number of surveyed patients}
\item{length_of_stay}{Average length of hospital stay}
\item{hospital_discharges}{Matrix of hospital discharge data (list-column)}
\item{population}{Population size}
\item{country}{Country code or name}
\item{loi_pps}{Length of infection PPS data (list-column)}
\item{num_hai_patients_by_stratum}{HAI patients by stratum (list-column)}
\item{mccabe_scores_distr}{Distribution of McCabe scores (list-column)}
\item{mccabe_life_exp}{Life expectancy by McCabe score (list-column)}
\item{num_hai_patients_by_stratum_prior}{Prior for HAI by stratum (list-column)}
\item{mccabe_by_stratum_prior}{Prior for McCabe by stratum (list-column)}
\item{num_survey_patients_by_stratum}{Survey patients by stratum (list-column)}
\item{bhai_options}{BHAI options used (list-column)}
\item{bhai_summary}{Summary statistics (list-column)}
}
}
\description{
Converts nested PPS (Point Prevalence Survey) S4 objects from BHAI
into a flat tibble format suitable for analysis and visualization.
}
\examples{
\dontrun{
  # Flatten German PPS data
  german_df <- flatten_pps(BHAI::german_pps_conv)
  head(german_df)

  # Flatten EU PPS data
  eu_df <- flatten_pps(BHAI::eu_pps)
  head(eu_df)
}
}
