% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthburdenr-package.R
\docType{package}
\name{healthburdenr-package}
\alias{healthburdenr}
\alias{healthburdenr-package}
\title{healthburdenr: Interactive Dashboard Comparing the Estimation of Burden of Healthcare-Associated Infections in Germany and European Union}
\description{
Provides tools and interactive visualizations for exploring healthcare-associated infection (HAI) burden estimates based on point prevalence survey data from European acute-care hospitals (2011-2012). Includes population estimates, stratified analyses, and disability-adjusted life years (DALYs) calculations. Based on Zacher et al.(2019)
}
\author{
\strong{Maintainer}: Jay Ashok Chauhan \email{jcha0338@student.monash.edu}

}
\keyword{internal}
