% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bhai_strata_summary}
\alias{bhai_strata_summary}
\title{BHAI Stratified Summary Results}
\format{
A data frame with stratified results:
\describe{
\item{country}{Country or region name}
\item{infection}{Type of healthcare-associated infection}
\item{sex}{Sex category (male/female)}
\item{age_group}{Age group category (factor)}
\item{ncases_lower_ci}{Lower 95\% CI for number of cases}
\item{ncases_estimate}{Median estimate for number of cases}
\item{ncases_upper_ci}{Upper 95\% CI for number of cases}
\item{ndeath_lower_ci}{Lower 95\% CI for number of deaths}
\item{ndeath_estimate}{Median estimate for number of deaths}
\item{ndeath_upper_ci}{Upper 95\% CI for number of deaths}
\item{daly_lower_ci}{Lower 95\% CI for DALYs}
\item{daly_estimate}{Median estimate for DALYs}
\item{daly_upper_ci}{Upper 95\% CI for DALYs}
}
}
\source{
Derived from BHAI simulations (nsim = 500) on German and EU PPS data
}
\usage{
bhai_strata_summary
}
\description{
Detailed stratified results from BHAI (Burden of Healthcare-Associated Infections)
analysis, including cases, deaths, and DALYs by country, infection type,
sex, and age group, with 95\% confidence intervals.
}
\examples{
data(bhai_strata_summary)

# View age groups
levels(bhai_strata_summary$age_group)

}
\seealso{
\link{bhai_pps_sample_distribution}, \link{bhai_pop_est}
}
\keyword{datasets}
